+++ src/cryptoexamples/example_asymmetric_string_encryption.py
@@ -34,7 +34,7 @@ def demonstrate_asymmetric_string_encryption(plain_text):
         cipher_text_bytes = public_key.encrypt(
             plaintext=plain_text.encode('utf-8'),
             padding=padding.OAEP(
-                mgf=padding.MGF1(algorithm=hashes.SHA256()),
+                mgf=padding.MGF1(algorithm=hashes.SHA512()),
                 algorithm=hashes.SHA512(),
                 label=None
             )
@@ -46,7 +46,7 @@ def demonstrate_asymmetric_string_encryption(plain_text):
         decrypted_cipher_text_bytes = private_key.decrypt(
             ciphertext=base64.urlsafe_b64decode(cipher_text),
             padding=padding.OAEP(
-                mgf=padding.MGF1(algorithm=hashes.SHA256()),
+                mgf=padding.MGF1(algorithm=hashes.SHA512()),
                 algorithm=hashes.SHA512(),
                 label=None
             )